requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
aiofiles
python-dotenv
aiohttp